requirements.txt
anthropic>=0.34.0
pdf2image>=1.16.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
python-pptx>=0.6.21
reportlab>=4.0
pyyaml>=6.0
cairosvg>=2.7.0
